+++ tools/cargo-do/src/main.rs
@@ -13,6 +13,7 @@ fn main() {
         "fmt" | "f" => fmt(args),
         "test" | "t" => test(args),
         "run" | "r" => run(args),
+        "zng" => cargo_zng(args),
         "doc" => doc(args),
         "expand" => expand(args),
         "check" | "c" => check(args),
@@ -459,7 +460,8 @@ fn test(mut args: Vec<&str>) {
 
         let all = args.is_empty();
 
-        if !all && args.contains(&"--doc") && !args.contains(&"-p") && !args.contains(&"--package") {
+        let has_package = args.contains(&"-p") || args.contains(&"--package");
+        if !all && args.contains(&"--doc") && !has_package {
             version_doc_sync::check();
         }
 
@@ -477,14 +479,27 @@ fn test(mut args: Vec<&str>) {
             }
             cmd_env(
                 "cargo",
-                &[nightly, "nextest", "run", "--no-fail-fast", "--all-features", "--workspace"],
+                &[
+                    nightly,
+                    "nextest",
+                    "run",
+                    "--no-fail-fast",
+                    "--all-features",
+                    if has_package { "" } else { "--workspace" },
+                ],
                 &args,
                 env,
             );
         } else {
             cmd_env(
                 "cargo",
-                &[nightly, "test", "--no-fail-fast", "--all-features", "--workspace"],
+                &[
+                    nightly,
+                    "test",
+                    "--no-fail-fast",
+                    "--all-features",
+                    if has_package { "" } else { "--workspace" },
+                ],
                 &args,
                 env,
             );
@@ -497,6 +512,12 @@ fn test(mut args: Vec<&str>) {
             test(vec!["--render"]);
         }
     }
+}
+
+// do zng [ARGS]
+//    Run cargo-zng with the same args for `cargo zng`.
+fn cargo_zng(args: Vec<&str>) {
+    cmd("cargo", &["run", "--package", "cargo-zng", "--", "zng"], &args);
 }
 
 // do run, r EXAMPLE [-b, --backtrace] [--release-lto] [<cargo-run-args>]